Handle errors in jupyterlab-lsp test
Turns out this was still failing because lots of http requests throw errors. Now we handle those errors, check that we got a response back, and that it is a 200 code.

I also noticed that the system command was returning the exit code, not the pid, so I fixed that.